+++ modules/admin/views/layouts/main.php
@@ -1,7 +1,6 @@
 <?php
 use \admin\Module;
 use \luya\Module as Luya;
-use \yii\helpers\Html;
 use \luya\helpers\Url;
 $user = Yii::$app->adminuser->getIdentity();
 $this->beginPage()
@@ -25,7 +24,6 @@ $this->beginPage()
         }
     </style>
     <?php $this->head(); ?>
-    <?= Html::csrfMetaTags() ?>
     <script>
         var authToken = '<?=$user->getAuthToken();?>';
     </script>